refactor(diff-schedule): split diff.ts into helpers and resolvers
diff.ts is down from 339 to 185 lines and only holds shared types plus
the computeDiff orchestrator. Pure utilities (toSlug, artistKey, date
math, timezone math) move to diffHelpers.ts. The per-row resolvers
(buildIndexes, resolveArtists, resolveStage, computeTimes, findMatchingSet)
move to diffResolvers.ts. diff.test.ts imports the helpers directly.
