+++ supabase/functions/diff-schedule/diff.test.ts
@@ -1,14 +1,16 @@
 import { assertEquals } from "jsr:@std/assert@1";
 import {
-  advanceDateByOne,
-  artistKey,
   computeDiff,
-  localToUtc,
-  toSlug,
   type DbArtist,
   type DbSet,
   type DbStage,
 } from "./diff.ts";
+import {
+  advanceDateByOne,
+  artistKey,
+  localToUtc,
+  toSlug,
+} from "./diffHelpers.ts";
 
 Deno.test("toSlug converts name to lowercase hyphenated slug", () => {
   assertEquals(toSlug("Carl Cox"), "carl-cox");
@@ -103,16 +105,19 @@ Deno.test("computeDiff: existing artist is not duplicated", () => {
   assertEquals(result.summary.newArtists, 0);
 });
 
-Deno.test("computeDiff: same new artist in multiple rows is created once", () => {
-  const result = computeDiff(
-    [{ artists: ["New DJ"] }, { artists: ["New DJ"] }],
-    [],
-    [],
-    [],
-    "Europe/Lisbon",
-  );
-  assertEquals(result.cleanOperations.artistsToCreate.length, 1);
-});
+Deno.test(
+  "computeDiff: same new artist in multiple rows is created once",
+  () => {
+    const result = computeDiff(
+      [{ artists: ["New DJ"] }, { artists: ["New DJ"] }],
+      [],
+      [],
+      [],
+      "Europe/Lisbon",
+    );
+    assertEquals(result.cleanOperations.artistsToCreate.length, 1);
+  },
+);
 
 Deno.test("computeDiff: CSV row with no DB match creates new set", () => {
   const result = computeDiff(
@@ -146,13 +151,7 @@ Deno.test("computeDiff: CSV row matching existing set produces update", () => {
 Deno.test("computeDiff: set in DB but absent from CSV is orphaned", () => {
   const artist = makeArtist("DJ Tennis");
   const set = makeSet("set-2", "DJ Tennis", [artist]);
-  const result = computeDiff(
-    [],
-    [],
-    [set],
-    [artist],
-    "Europe/Lisbon",
-  );
+  const result = computeDiff([], [], [set], [artist], "Europe/Lisbon");
   assertEquals(result.conflicts.orphanedSets.length, 1);
   assertEquals(result.conflicts.orphanedSets[0].id, "set-2");
   assertEquals(result.summary.setsOrphaned, 1);
@@ -187,18 +186,24 @@ Deno.test("computeDiff: B2B artist order in CSV does not affect match", () => {
   assertEquals(result.cleanOperations.setsToUpdate.length, 1);
 });
 
-Deno.test("computeDiff: exact stage name match uses canonical DB name in payload", () => {
-  const artist = makeArtist("Carl Cox");
-  const stage = makeStage("stage-1", "Main Stage");
-  const result = computeDiff(
-    [{ artists: ["Carl Cox"], stage: "Main Stage" }],
-    [stage],
-    [],
-    [artist],
-    "Europe/Lisbon",
-  );
-  assertEquals(result.cleanOperations.setsToCreate[0].stageName, "Main Stage");
-});
+Deno.test(
+  "computeDiff: exact stage name match uses canonical DB name in payload",
+  () => {
+    const artist = makeArtist("Carl Cox");
+    const stage = makeStage("stage-1", "Main Stage");
+    const result = computeDiff(
+      [{ artists: ["Carl Cox"], stage: "Main Stage" }],
+      [stage],
+      [],
+      [artist],
+      "Europe/Lisbon",
+    );
+    assertEquals(
+      result.cleanOperations.setsToCreate[0].stageName,
+      "Main Stage",
+    );
+  },
+);
 
 Deno.test("computeDiff: stage name mismatch surfaced as conflict", () => {
   const artist = makeArtist("Carl Cox");
@@ -212,7 +217,10 @@ Deno.test("computeDiff: stage name mismatch surfaced as conflict", () => {
   );
   assertEquals(result.conflicts.stageNameMismatches.length, 1);
   assertEquals(result.conflicts.stageNameMismatches[0].csvValue, "Mainstage");
-  assertEquals(result.conflicts.stageNameMismatches[0].closestDbValue, "Main Stage");
+  assertEquals(
+    result.conflicts.stageNameMismatches[0].closestDbValue,
+    "Main Stage",
+  );
 });
 
 Deno.test("computeDiff: unknown stage creates new stage", () => {
@@ -228,62 +236,84 @@ Deno.test("computeDiff: unknown stage creates new stage", () => {
   assertEquals(result.cleanOperations.stagesToCreate[0].name, "Secret Forest");
 });
 
-Deno.test("computeDiff: end time before start time triggers midnight advance", () => {
-  const artist = makeArtist("Carl Cox");
-  const result = computeDiff(
-    [{ artists: ["Carl Cox"], date: "2026-07-11", startTime: "23:00", endTime: "01:00" }],
-    [],
-    [],
-    [artist],
-    "UTC",
-  );
-  const created = result.cleanOperations.setsToCreate[0];
-  // start should be 2026-07-11T23:00:00Z, end should be 2026-07-12T01:00:00Z
-  assertEquals(created.timeStart, "2026-07-11T23:00:00.000Z");
-  assertEquals(created.timeEnd, "2026-07-12T01:00:00.000Z");
-});
+Deno.test(
+  "computeDiff: end time before start time triggers midnight advance",
+  () => {
+    const artist = makeArtist("Carl Cox");
+    const result = computeDiff(
+      [
+        {
+          artists: ["Carl Cox"],
+          date: "2026-07-11",
+          startTime: "23:00",
+          endTime: "01:00",
+        },
+      ],
+      [],
+      [],
+      [artist],
+      "UTC",
+    );
+    const created = result.cleanOperations.setsToCreate[0];
+    // start should be 2026-07-11T23:00:00Z, end should be 2026-07-12T01:00:00Z
+    assertEquals(created.timeStart, "2026-07-11T23:00:00.000Z");
+    assertEquals(created.timeEnd, "2026-07-12T01:00:00.000Z");
+  },
+);
 
-Deno.test("computeDiff: set name falls back to b2b join when not provided", () => {
-  const artist1 = makeArtist("Carl Cox");
-  const artist2 = makeArtist("Peggy Gou");
-  const result = computeDiff(
-    [{ artists: ["Carl Cox", "Peggy Gou"] }],
-    [],
-    [],
-    [artist1, artist2],
-    "UTC",
-  );
-  assertEquals(result.cleanOperations.setsToCreate[0].name, "Carl Cox b2b Peggy Gou");
-});
+Deno.test(
+  "computeDiff: set name falls back to b2b join when not provided",
+  () => {
+    const artist1 = makeArtist("Carl Cox");
+    const artist2 = makeArtist("Peggy Gou");
+    const result = computeDiff(
+      [{ artists: ["Carl Cox", "Peggy Gou"] }],
+      [],
+      [],
+      [artist1, artist2],
+      "UTC",
+    );
+    assertEquals(
+      result.cleanOperations.setsToCreate[0].name,
+      "Carl Cox b2b Peggy Gou",
+    );
+  },
+);
 
-Deno.test("computeDiff: explicit set name takes precedence over b2b fallback", () => {
-  const artist = makeArtist("Carl Cox");
-  const result = computeDiff(
-    [{ artists: ["Carl Cox"], setName: "Carl Cox Live" }],
-    [],
-    [],
-    [artist],
-    "UTC",
-  );
-  assertEquals(result.cleanOperations.setsToCreate[0].name, "Carl Cox Live");
-});
+Deno.test(
+  "computeDiff: explicit set name takes precedence over b2b fallback",
+  () => {
+    const artist = makeArtist("Carl Cox");
+    const result = computeDiff(
+      [{ artists: ["Carl Cox"], setName: "Carl Cox Live" }],
+      [],
+      [],
+      [artist],
+      "UTC",
+    );
+    assertEquals(result.cleanOperations.setsToCreate[0].name, "Carl Cox Live");
+  },
+);
 
-Deno.test("computeDiff: same stage mismatch from multiple rows surfaced once", () => {
-  const artist1 = makeArtist("Artist A");
-  const artist2 = makeArtist("Artist B");
-  const stage = makeStage("stage-1", "Main Stage");
-  const result = computeDiff(
-    [
-      { artists: ["Artist A"], stage: "Mainstage" },
-      { artists: ["Artist B"], stage: "Mainstage" },
-    ],
-    [stage],
-    [],
-    [artist1, artist2],
-    "UTC",
-  );
-  assertEquals(result.conflicts.stageNameMismatches.length, 1);
-});
+Deno.test(
+  "computeDiff: same stage mismatch from multiple rows surfaced once",
+  () => {
+    const artist1 = makeArtist("Artist A");
+    const artist2 = makeArtist("Artist B");
+    const stage = makeStage("stage-1", "Main Stage");
+    const result = computeDiff(
+      [
+        { artists: ["Artist A"], stage: "Mainstage" },
+        { artists: ["Artist B"], stage: "Mainstage" },
+      ],
+      [stage],
+      [],
+      [artist1, artist2],
+      "UTC",
+    );
+    assertEquals(result.conflicts.stageNameMismatches.length, 1);
+  },
+);
 
 Deno.test("computeDiff: multiple candidates disambiguated by stage", () => {
   const artist = makeArtist("Carl Cox");
